+++ packages/talker_logger/lib/src/src.dart
@@ -1,4 +1,3 @@
-export 'extensions/extensions.dart';
 export 'models/models.dart';
 export 'settings/settings.dart';
 export 'talker_logger.dart';