+++ packages/talker_logger/lib/src/utils/console_utils.dart
@@ -1,0 +1,9 @@
+/// Сlass that contains methods for working with console output
+abstract class ConsoleUtils {
+  ConsoleUtils._();
+
+  /// Method returns a line for the bottom of the message
+  static String getUnderline(int length, {String lineSymbol = '-'}) {
+    return lineSymbol * length;
+  }
+}