+++ packages/talker_logger/lib/src/utils/filter/loglevel_logger_filter.dart
@@ -1,7 +1,10 @@
 import 'package:talker_logger/talker_logger.dart';
 
-class BaseTalkerLoggerFilter implements TalkerLoggerFilter {
-  const BaseTalkerLoggerFilter(this.logLevel);
+/// This filter checks that current message level
+/// is above certain [LogLevel] setting in [TalkerLoggerSettings]
+class LogLevelTalkerLoggerFilter implements TalkerLoggerFilter {
+  const LogLevelTalkerLoggerFilter(this.logLevel);
+
   final LogLevel logLevel;
 
   @override