+++ packages/talker_logger/lib/src/utils/filter/talker_logger_filter.dart
@@ -1,5 +1,0 @@
-import 'package:talker_logger/talker_logger.dart';
-
-abstract class TalkerLoggerFilter {
-  bool shouldLog(String msg, LogLevel level);
-}